requirements.txt
torch==2.0.1
torchvision==0.15.2
numpy==1.24.3
opencv-python==4.8.0.74
scikit-image==0.21.0
tensorboard==2.13.0
matplotlib==3.7.1
scipy==1.10.1
tqdm==4.65.0
pillow==9.5.0
pyyaml==6.0